Update to support TLS security
Included a configuration file to use that configures apache to use TLS
security for the Nextcloud site. This normally works on the first run
because the install script creates this file but on reboot the file
reverts back to the default for the container. Will put this file in the
container in future.