Install and setup tzdata separately
Previously the istallation would pause during the installation of
mariadb with an interactive prompt asking the user for timezone
information. This prompt was coming from one of the dependencies of
mariadb called tzdata which shares timezone information about the system
to multiple applications. To make the installation headless, tzdata is
now installed seprately with that prompt suppressed and the timezone set
to UCT be default or the value of the TIME_ZONE environment variable.