Change apache configuration to use http instead of https
The final config will use traefik as reverse proxy to handle certificate
requests and rotation. So the Nextcloud server should not need to use
https.

We may need to reverse this change in the future if Nextcloud still
disables some functionality because it won't know that there is a
reverse proxy applying encryption on its behalf.