Switch off skip-name-resolve on mariadb
MariaDB was not resolving the container names used in the database user
names. The requests that came from the NextCloud container were seen as
coming from the containers ip address and thus being blocked by the
database server because it wasn't resolving the hostname in the username
and seeing that the ip addresses match.

This is an option of mariadb called skip-name-resolve which was switched
on by the recommended config from Nextcloud. I think this is a security
thing because a DNS resolution of a hostname can be spoofed by a rouge
DNS server. Forcing the server to use the IP addresses is thus safer. I
think the Nextcloud team assumed that the database would run on the same
host as the installation but that's not the case here. Being in a
containerised environment means I don't really have to worry about DNS
spoofing. Name resolution is done by the docker engine.